+++ HISTORY.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.2] - 2023-09-18
+
 ### Added
 
 - `onAfterInteractive`, `onBeforeInteractive`, `onUnsupported` callbacks
@@ -118,7 +120,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial release.
 
-[unreleased]: https://github.com/Le0Developer/react-turnstile/compare/v1.1.1...HEAD
+[unreleased]: https://github.com/Le0Developer/react-turnstile/compare/v1.1.2...HEAD
+[1.1.2]: https://github.com/le0developer/react-turnstile/compare/v1.1.1...v1.1.2
 [1.1.1]: https://github.com/le0developer/react-turnstile/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/le0developer/react-turnstile/compare/v1.0.6...v1.1.0
 [1.0.6]: https://github.com/le0developer/react-turnstile/compare/v1.0.5...v1.0.6